Adopt DesignScaffoldControls.LabeledSlider — and drop a truncating Int bridge
`ParameterSlider` was one of SEVEN copies of this row on the volume, promoted into
DesignScaffold 0.16.0. Five call sites move over.

Two of them were wrapping an `Int` parameter in a hand-written `Binding<Double>` with
`set: { parameters.topK = Int($0) }` — which TRUNCATES. `Int(511.9999)` is 511, so the
top of a drag quietly lost a unit; SenseNova's copy of the same wrapper rounded, and one
of the two had to be wrong. `LabeledSlider` takes the `Int` binding directly and rounds,
so `topKSliderRange` and `maxFramesSliderRange` are now `ClosedRange<Int>` — the Double
versions only ever existed because the old row could not take anything else.

`help:` is gone as a parameter: the row applies `.help(_:)` like any other modifier, which
also retires `.help(help ?? "")` attaching an empty tooltip to every row without one.

Visible change: readouts now format through `FormatStyle` rather than `"%.0f"`, so Max
frames reads `1,024` — grouped and locale-aware — where it read `1024`. Verified running.

+++ Audio8 Demo/Model/RunMetrics.swift
@@ -59,8 +59,11 @@ struct RunParameters: Codable, Hashable, Sendable {
     static let maxCharacters = 1000
 
     /// Slider-facing mirrors of the integer ranges (SwiftUI sliders are Double-valued).
-    static let topKSliderRange: ClosedRange<Double> = 1...200
-    static let maxFramesSliderRange: ClosedRange<Double> = 32...2048
+    // Int, matching the parameters they bound: `LabeledSlider` takes an `Int` binding
+    // directly, so nothing here has to convert. The Double versions existed only because
+    // the old hand-rolled slider could not.
+    static let topKSliderRange: ClosedRange<Int> = 1...200
+    static let maxFramesSliderRange: ClosedRange<Int> = 32...2048
 }
 
 /// How the reference voice was supplied.
